requirements.txt
black
apache-airflow[postgres,google]==2.3.0
pendulum
pytest
pandas